requirements.txt
flask==2.0.1
werkzeug==2.0.1
sqlalchemy==1.4.23
flask-sqlalchemy==2.5.1
flask-restx==0.5.1
flask-cors==3.0.10
flask-limiter==2.4.0
python-dotenv==0.19.0
pyjwt==2.1.0
bcrypt==3.2.0
email-validator==1.1.3
pytest==7.0.0